% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{calc_ndvi}
\alias{calc_ndvi}
\title{Calcul du NDVI (Normalized Difference Vegetation Index)}
\usage{
calc_ndvi(nir, red)
}
\arguments{
\item{nir}{Valeur de la bande proche infrarouge (0-1 ou 0-255)}

\item{red}{Valeur de la bande rouge (0-1 ou 0-255)}
}
\value{
Valeur NDVI entre -1 et 1
}
\description{
Calcule l'indice de végétation par différence normalisée à partir
des bandes spectrales NIR et Rouge.
}
\details{
Le NDVI est calculé selon la formule :
\deqn{NDVI = \frac{NIR - Rouge}{NIR + Rouge}}

Interprétation :
\itemize{
\item NDVI < 0 : eau, neige, nuages
\item 0.0 - 0.2 : sol nu, zones urbanisées
\item 0.2 - 0.4 : végétation éparse
\item 0.4 - 0.6 : végétation modérée (cultures)
\item 0.6 - 1.0 : végétation dense et saine
}
}
\examples{
# NDVI d'un pixel de culture saine
calc_ndvi(nir = 0.8, red = 0.1)

# Sur un vecteur de pixels
nir_vals <- c(0.8, 0.3, 0.05, 0.6)
red_vals <- c(0.1, 0.2, 0.04, 0.15)
calc_ndvi(nir_vals, red_vals)

}
