% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_biomass.R
\name{calculate_biomass}
\alias{calculate_biomass}
\title{Estimate vegetation biomass}
\usage{
calculate_biomass(ndvi, lai = NULL, method = "linear")
}
\arguments{
\item{ndvi}{Raster ou valeurs NDVI}

\item{lai}{Raster ou valeurs LAI optionnel}

\item{method}{Methode d'estimation}
}
\value{
Biomasse estimee (kg/ha)
}
\description{
Estime la biomasse vegetale disponible
a partir du NDVI et du LAI.
}
\details{
Relation simplifiee :

Biomass = a + b * NDVI

Par defaut :

Biomass = 1000 * NDVI

Si LAI disponible :

Biomass = 800 * NDVI + 200 * LAI

Compatible avec les rasters terra.
}
\examples{
biomass <- calculate_biomass(
  ndvi = 0.6
)

}
\references{
Tucker, C.J. (1979).
Red and photographic infrared linear combinations
for monitoring vegetation.
Remote Sensing of Environment.
}
