% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grazing_pressure.R
\name{calculate_grazing_pressure}
\alias{calculate_grazing_pressure}
\title{Calcul de la pression pastorale}
\usage{
calculate_grazing_pressure(biomass, carrying_capacity, animal_density = NULL)
}
\arguments{
\item{biomass}{Biomasse disponible (kg/ha).
Numerique ou SpatRaster.}

\item{carrying_capacity}{Capacite de charge estimee
(UBT/ha). Numerique ou SpatRaster.}

\item{animal_density}{Densite animale reelle
(UBT/ha). Optionnelle.}
}
\value{
Indice de pression pastorale.
}
\description{
Analyse la pression de paturage en comparant
la densite animale avec la capacite de charge disponible.
}
\details{
La pression pastorale est calculee par :

\deqn{
GrazingPressure =
\frac{AnimalDensity}{CarryingCapacity}
}

Interpretation :
\itemize{
\item inferieur a 1 : faible pression (sous-paturage)
\item egal a 1 : equilibre
\item superieur a 1 : forte pression (surpaturage)
}

Si aucune densite animale n'est fournie,
la fonction utilise la biomasse disponible
pour produire un indice relatif normalise
entre 0 et 1.

Compatible avec les rasters terra.
}
\examples{

pressure <- calculate_grazing_pressure(
  biomass = 500,
  carrying_capacity = 2,
  animal_density = 1
)

}
\references{
Holechek, J.L., Pieper, R.D.,
& Herbel, C.H. (2010).
Range Management:
Principles and Practices.
}
