% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ndvi.R
\name{calculate_ndvi}
\alias{calculate_ndvi}
\title{Calcul du NDVI (Normalized Difference Vegetation Index)}
\usage{
calculate_ndvi(nir, red)
}
\arguments{
\item{nir}{Bande proche infrarouge}

\item{red}{Bande rouge}
}
\value{
Valeurs ou raster NDVI
}
\description{
Calcule le NDVI à partir des bandes NIR et Rouge.
}
\details{
Le NDVI est calculé selon :

\deqn{
NDVI = \frac{NIR - Red}{NIR + Red}
}

Les valeurs sont généralement comprises
entre -1 et 1.
}
\examples{
calculate_ndvi(0.8, 0.1)

}
\references{
Tucker, C.J. (1979).
Red and photographic infrared linear combinations
for monitoring vegetation.
}
