% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{classifier_ndvi}
\alias{classifier_ndvi}
\title{Classifier le NDVI en catégories de végétation}
\usage{
classifier_ndvi(ndvi)
}
\arguments{
\item{ndvi}{Vecteur de valeurs NDVI}
}
\value{
Facteur avec les classes de végétation
}
\description{
Classifier le NDVI en catégories de végétation
}
\examples{
valeurs <- c(-0.1, 0.1, 0.3, 0.5, 0.75)
classifier_ndvi(valeurs)

}
