% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{classify_biomass}
\alias{classify_biomass}
\title{Classification de la biomasse pastorale}
\usage{
classify_biomass(biomass)
}
\arguments{
\item{biomass}{Valeurs de biomasse en kg/ha}
}
\value{
Facteur contenant les classes de biomasse
}
\description{
Classe les valeurs de biomasse en catégories pastorales.
}
\details{
Les classes utilisées sont :
\itemize{
\item Très faible
\item Faible
\item Modérée
\item Élevée
\item Très élevée
}
}
\examples{
biomasse <- c(5, 20, 45, 70, 95)

classify_biomass(biomasse)

}
