% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrying_capacity.R
\name{classify_carrying_capacity}
\alias{classify_carrying_capacity}
\title{Classification de la capacité pastorale}
\usage{
classify_carrying_capacity(capacity)
}
\arguments{
\item{capacity}{Valeurs de capacité pastorale}
}
\value{
Facteur contenant les classes pastorales
}
\description{
Classe les valeurs de capacité de charge en catégories.
}
\examples{
values <- c(0.2, 0.8, 1.5, 3, 5)

classify_carrying_capacity(values)

}
