% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_occurrences.R
\name{clean_occurrences}
\alias{clean_occurrences}
\title{Clean species occurrences}
\usage{
clean_occurrences(occurrences)
}
\arguments{
\item{occurrences}{Objet sf contenant les occurrences}
}
\value{
Objet sf nettoye
}
\description{
Nettoie les donnees spatiales d'occurrences.
}
\details{
La fonction realise :
\itemize{
\item suppression des coordonnees invalides
\item suppression des doublons spatiaux
\item filtrage spatial simple
}
}
\examples{
\dontrun{
clean <- clean_occurrences(species$sf_object)
}

}
