% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrying_capacity.R
\name{estimate_carrying_capacity}
\alias{estimate_carrying_capacity}
\title{Estimation de la capacité de charge pastorale}
\usage{
estimate_carrying_capacity(biomass, consumption)
}
\arguments{
\item{biomass}{Biomasse disponible en kg/ha}

\item{consumption}{Consommation moyenne d'un animal en kg}
}
\value{
Capacité de charge animale
}
\description{
Estime le nombre d'animaux qu'une zone pastorale
peut supporter à partir de la biomasse disponible.
}
\details{
La capacité de charge est calculée selon :
\deqn{
Capacite = \frac{Biomasse}{Consommation animale}
}

Une capacité élevée indique une bonne disponibilité
fourragère et une pression pastorale faible.
}
\examples{
# Biomasse disponible
biomass <- 500

# Consommation moyenne d'un animal
consumption <- 250

estimate_carrying_capacity(biomass, consumption)

}
\references{
Holechek, J.L., Pieper, R.D., & Herbel, C.H. (2010).
Range Management: Principles and Practices.
}
