% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluation du modèle SDM}
\usage{
evaluate_model(observed, predicted)
}
\arguments{
\item{observed}{Valeurs observées}

\item{predicted}{Valeurs prédites}
}
\value{
Dataframe contenant Accuracy et RMSE
}
\description{
Calcule les performances du modèle de distribution
des espèces fourragères.
}
\details{
Les métriques calculées sont :
\itemize{
\item Accuracy
\item RMSE
}
}
\examples{
observed <- c(1,0,1,1,0)
predicted <- c(1,0,1,0,0)

evaluate_model(observed, predicted)

}
\references{
Fielding, A.H. & Bell, J.F. (1997).
A review of methods for the assessment
of prediction errors in conservation presence/absence models.
}
