% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_climate_data.R
\name{load_climate_data}
\alias{load_climate_data}
\title{Import climate data from WorldClim}
\usage{
load_climate_data(path = "climate_data", resolution = 10)
}
\arguments{
\item{path}{Dossier de stockage des donnees
climatiques.}

\item{resolution}{Resolution spatiale
WorldClim en minutes.}
}
\value{
Raster climatique de type SpatRaster.
}
\description{
Import des donnees climatiques pour
la modelisation des especes fourrageres.
}
\details{
Cette fonction telecharge les variables
bioclimatiques depuis WorldClim.

Variables incluses :
\itemize{
\item temperature,
\item precipitations.
}

Les donnees sont utilisees dans :
\itemize{
\item les SDM,
\item la modelisation spatiale,
\item l'analyse environnementale.
}

Fonctionnalites :
\itemize{
\item telechargement automatique,
\item raster stacking,
\item harmonisation spatiale.
}

Source :
WorldClim database.
}
\examples{
\dontrun{
climate <- load_climate_data()
plot(climate[[1]])
}

}
\references{
Fick, S.E. & Hijmans, R.J. (2017).
WorldClim 2: new climate surfaces for global land areas.
International Journal of Climatology.

WorldClim:
https://www.worldclim.org/
}
