% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sdm_model.R
\name{train_sdm_model}
\alias{train_sdm_model}
\title{Train Species Distribution Model (SDM)}
\usage{
train_sdm_model(
  data,
  target = "presence",
  predictors_raster = NULL,
  ntree_values = c(100, 300, 500)
)
}
\arguments{
\item{data}{Dataframe contenant les variables environnementales
et la variable présence.}

\item{target}{Variable cible.}

\item{predictors_raster}{Raster environnemental optionnel.}

\item{ntree_values}{Valeurs de ntree à tester.}
}
\value{
Liste contenant :
\itemize{
\item modèle SDM
\item importance variables
\item performances train/test
\item suitability raster (optionnel)
}
}
\description{
Modélisation des espèces fourragères.
}
\examples{
\dontrun{
model <- train_sdm_model(
  data,
  target = "presence"
)
}

}
