% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{calculate_biomass}
\alias{calculate_biomass}
\title{Estimation simplifiée de la biomasse à partir du NDVI}
\usage{
calculate_biomass(ndvi)
}
\arguments{
\item{ndvi}{Valeurs NDVI comprises entre -1 et 1}
}
\value{
Biomasse estimée en kg/ha
}
\description{
Estime la biomasse végétale aérienne à partir des valeurs NDVI
en utilisant une relation linéaire simplifiée inspirée des études
de télédétection appliquées aux écosystèmes pastoraux.
}
\details{
La biomasse est estimée selon la relation :
\deqn{Biomasse = 100 \times NDVI}

Cette relation simplifiée suppose qu'une augmentation du NDVI
correspond à une augmentation de la densité et de la productivité
de la végétation.

Interprétation :
\itemize{
\item Biomasse faible : végétation sparse ou zones dégradées
\item Biomasse modérée : parcours exploitables
\item Biomasse élevée : forte disponibilité fourragère
}
}
\examples{
# Biomasse d'une végétation dense
calculate_biomass(0.75)

# Biomasse sur plusieurs pixels
ndvi_vals <- c(0.1, 0.3, 0.5, 0.8)

calculate_biomass(ndvi_vals)

}
\references{
Tucker, C.J. (1979).
Red and photographic infrared linear combinations
for monitoring vegetation.
Remote Sensing of Environment, 8(2), 127-150.

Paruelo, J.M., Epstein, H.E., Lauenroth, W.K.,
& Burke, I.C. (1997).
ANPP estimates from NDVI for the central grassland
region of the United States.
Ecology, 78(3), 953-958.
}
