% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grazing_pressure.R
\name{calculate_grazing_pressure}
\alias{calculate_grazing_pressure}
\title{Calcul de la pression pastorale}
\usage{
calculate_grazing_pressure(biomass, carrying_capacity, animal_density = NULL)
}
\arguments{
\item{biomass}{Biomasse disponible (kg/ha).}

\item{carrying_capacity}{Capacité de charge estimée
(UBT/ha).}

\item{animal_density}{Densité animale réelle
(UBT/ha). Optionnelle.}
}
\value{
Indice de pression pastorale.
}
\description{
Analyse la pression de pâturage en comparant
la densité animale avec la capacité de charge disponible.
}
\details{
La pression pastorale est calculée par :

\deqn{
GrazingPressure =
\frac{AnimalDensity}{CarryingCapacity}
}

Interprétation :
\itemize{
\item inférieur à 1 : faible pression (sous-pâturage)
\item égal à 1 : équilibre
\item supérieur à 1 : forte pression (surpâturage)
}

Si aucune densité animale n'est fournie,
la fonction utilise la biomasse disponible
pour produire un indice relatif.

Compatible avec les rasters terra.
}
\examples{

pressure <- calculate_grazing_pressure(
  biomass = 500,
  carrying_capacity = 2,
  animal_density = 1
)

}
\references{
Holechek, J.L., Pieper, R.D.,
& Herbel, C.H. (2010).
Range Management:
Principles and Practices.
}
