% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ndvi.R
\name{calculate_ndvi}
\alias{calculate_ndvi}
\title{Calcul du NDVI (Normalized Difference Vegetation Index)}
\usage{
calculate_ndvi(nir, red)
}
\arguments{
\item{nir}{Bande proche infrarouge (valeur numerique ou SpatRaster).}

\item{red}{Bande rouge (valeur numerique ou SpatRaster).}
}
\value{
Valeurs ou raster NDVI compris entre -1 et 1.
}
\description{
Calcule le NDVI a partir des bandes NIR et Rouge.
}
\details{
Le NDVI est calcule selon :

\deqn{
NDVI = \frac{NIR - Red}{NIR + Red}
}

Les pixels ou NIR + Red = 0 produisent NaN
(comportement standard conforme a la norme NDVI).

Compatible avec les scalaires, vecteurs et SpatRaster.
}
\examples{
calculate_ndvi(0.8, 0.1)

}
\references{
Tucker, C.J. (1979).
Red and photographic infrared linear combinations
for monitoring vegetation.
Remote Sensing of Environment.
}
