% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ndvi_lai.R
\name{download_ndvi_lai}
\alias{download_ndvi_lai}
\title{Download NDVI and LAI data}
\usage{
download_ndvi_lai(ndvi_file = NULL, lai_file = NULL)
}
\arguments{
\item{ndvi_file}{Chemin raster NDVI optionnel}

\item{lai_file}{Chemin raster LAI optionnel}
}
\value{
Liste contenant les rasters NDVI et LAI
}
\description{
Telecharge ou importe les donnees NDVI et LAI.
}
\details{
Fonctionnalites :
\itemize{
\item chargement raster NDVI
\item chargement raster LAI
\item preparation pour analyse spatiale
}
}
\examples{
veg <- download_ndvi_lai()

}
