% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_carrying_capacity.R
\name{estimate_carrying_capacity}
\alias{estimate_carrying_capacity}
\title{Estimation de la capacité de charge pastorale}
\usage{
estimate_carrying_capacity(biomass, animal_demand = 250)
}
\arguments{
\item{biomass}{Biomasse disponible en kg/ha.
Peut être une valeur numérique ou un raster terra.}

\item{animal_demand}{Consommation moyenne d'une unité animale
en kg de matière sèche.}
}
\value{
Liste contenant :
\itemize{
\item capacity : capacité de charge (UBT/ha)
\item pressure : classe de pression pastorale
}
}
\description{
Estime le nombre d'unités animales qu'une zone pastorale
peut supporter à partir de la biomasse disponible.
}
\details{
La capacité de charge est calculée selon :

\deqn{
CarryingCapacity =
\frac{AvailableBiomass}{AnimalDemand}
}

Classification :
\itemize{
\item Faible : < 0.5 UBT/ha
\item Modérée : 0.5 - 1 UBT/ha
\item Elevée : > 1 UBT/ha
}
}
\examples{
capacity <- estimate_carrying_capacity(
  biomass = 500,
  animal_demand = 250
)

}
\references{
Holechek, J.L., Pieper, R.D.,
& Herbel, C.H. (2010).
Range Management:
Principles and Practices.
}
