% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_refuge_areas.R
\name{identify_refuge_areas}
\alias{identify_refuge_areas}
\title{Identification des zones refuges pastorales}
\usage{
identify_refuge_areas(biomass, grazing_pressure, climate_score = 1)
}
\arguments{
\item{biomass}{Raster ou valeurs de biomasse disponible.}

\item{grazing_pressure}{Raster ou valeurs d'indice
de pression pastorale.}

\item{climate_score}{Conditions climatiques favorables.
Valeurs comprises entre 0 et 1.}
}
\value{
Liste contenant :
\itemize{
\item refuge_index
\item priority_area
}
}
\description{
Identifie les zones prioritaires pour la conservation
en combinant biomasse, pression pastorale et climat.
}
\details{
Une zone refuge correspond à :
\itemize{
\item biomasse élevée,
\item faible pression pastorale,
\item conditions climatiques favorables.
}

L'indice est calculé par :

\deqn{
RefugeIndex =
BiomassScore *
ClimateScore /
GrazingPressure
}

Classification :
\itemize{
\item FALSE : zone non prioritaire
\item TRUE : zone refuge prioritaire
}

Compatible avec les rasters terra.
}
\examples{

refuge <- identify_refuge_areas(
  biomass = 800,
  grazing_pressure = 0.5,
  climate_score = 0.8
)

}
\references{
Franklin, J. (2010).
Mapping Species Distributions:
Spatial Inference and Prediction.
}
