% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_biomass_data.R
\name{import_biomass_data}
\alias{import_biomass_data}
\title{Import biomass field data}
\usage{
import_biomass_data(file)
}
\arguments{
\item{file}{Chemin du fichier CSV ou Excel}
}
\value{
Dataframe biomasse nettoye
}
\description{
Importe les donnees terrain de biomasse.
}
\details{
Fonctionnalites :
\itemize{
\item import CSV / Excel
\item gestion des valeurs manquantes
\item harmonisation des unites
}

Variables attendues :
\itemize{
\item plot
\item date
\item biomass
}
}
\examples{
\dontrun{
biomass <- import_biomass_data(
  "biomass.csv"
)
}

}
