% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_forage_species.R
\name{import_forage_species}
\alias{import_forage_species}
\title{Import forage species data}
\usage{
import_forage_species(species_name = NULL, csv_file = NULL, limit = 100)
}
\arguments{
\item{species_name}{Scientific species name.}

\item{csv_file}{CSV file path.}

\item{limit}{Number of GBIF occurrences.}
}
\value{
A list containing:
\itemize{
\item sf_object : spatial sf object
\item dataframe : species dataframe
}
}
\description{
Import forage species occurrences from
GBIF or field CSV files.
}
\details{
This function allows:
\itemize{
\item GBIF import,
\item field CSV import,
\item duplicate removal,
\item coordinate cleaning,
\item conversion to spatial sf object.
}

Variables:
\itemize{
\item species
\item GPS coordinates
\item optional abundance
}
}
\examples{
\dontrun{
species <- import_forage_species(
  species_name = "Stipa tenacissima"
)
}

}
\references{
GBIF.org (2025).
Global Biodiversity Information Facility.

Pebesma, E. (2018).
Simple Features for R.
}
