% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_predictors.R
\name{prepare_predictors}
\alias{prepare_predictors}
\title{Preparation des variables environnementales}
\usage{
prepare_predictors(occurrence, predictors, remove_cor = TRUE)
}
\arguments{
\item{occurrence}{Objet sf contenant les points
de presence des especes.}

\item{predictors}{Raster contenant les variables
environnementales (NDVI, LAI, climat, altitude).}

\item{remove_cor}{Logical. Supprimer ou non les
variables fortement correlees.}
}
\value{
Dataframe pret pour la modelisation ML.
}
\description{
Prepare les predicteurs environnementaux utilises
pour la modelisation des especes fourrageres (SDM).
}
\details{
Cette fonction permet :
\itemize{
\item extraction des valeurs raster aux points GPS,
\item combinaison presence + environnement,
\item suppression des valeurs manquantes,
\item suppression optionnelle des variables correlees.
}

Variables possibles :
\itemize{
\item NDVI,
\item LAI,
\item temperature,
\item precipitation,
\item altitude.
}

Les donnees produites sont utilisees par :
train_sdm_model().
}
\examples{

\dontrun{

data <- prepare_predictors(
  occurrence = species$sf_object,
  predictors = climate
)

}

}
\references{
Elith, J. & Leathwick, J.R. (2009).
Species Distribution Models:
Ecological explanation and prediction.
Annual Review of Ecology Evolution and Systematics.
}
