% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_model.R
\name{train_sdm_model}
\alias{train_sdm_model}
\title{Train Species Distribution Model}
\usage{
train_sdm_model(data)
}
\arguments{
\item{data}{Dataframe contenant les variables environnementales
et la colonne presence}
}
\value{
Modèle Random Forest entraîné
}
\description{
Entraîne un modèle Random Forest pour prédire
la présence des espèces fourragères.
}
\details{
Le modèle utilise :
\itemize{
\item NDVI
\item température
\item précipitations
pour prédire la présence des espèces.
}
}
\examples{
data <- data.frame(
  presence = c(1,0,1,0,1),
  ndvi = c(0.7,0.2,0.8,0.1,0.6),
  temperature = c(18,35,20,38,19),
  rainfall = c(400,90,350,70,420)
)

model <- train_sdm_model(data)

}
