% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_sdm_model.R
\name{train_sdm_model}
\alias{train_sdm_model}
\title{Train Species Distribution Model (SDM)}
\usage{
train_sdm_model(
  data,
  target = "presence",
  predictors_raster = NULL,
  ntree_values = c(100, 300, 500)
)
}
\arguments{
\item{data}{Dataframe contenant les variables environnementales
et la variable presence.}

\item{target}{Variable cible (presence/absence). Par defaut "presence".}

\item{predictors_raster}{Raster environnemental optionnel.
Si fourni, un raster de suitabilite est produit (probabilites).}

\item{ntree_values}{Valeurs de ntree a tester lors du tuning.}
}
\value{
Liste contenant :
\itemize{
\item model : meilleur modele Random Forest
\item best_ntree : valeur ntree selectionnee
\item train_accuracy : precision sur le jeu d'entrainement
\item test_accuracy : precision sur le jeu de test
\item importance : importance des variables
\item suitability_raster : raster de probabilites (si predictors_raster fourni)
}
}
\description{
Modelisation des especes fourrageres par Random Forest.
}
\details{
Le tuning selectionne le ntree maximisant l'accuracy
sur le jeu de test (30 pourcent des donnees).

Le raster de suitabilite contient les probabilites
de presence (classe "1") issues de predict(..., type = "prob").
}
\examples{
\dontrun{
model <- train_sdm_model(
  data,
  target = "presence"
)
}

}
